requirements.txt
pandas==2.2.2
pyyaml==6.0.2
pytest